= newrelic: cleanup
